requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
happybase==1.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
kafka-python==2.0.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
oauthlib==3.1.0
ply==3.11
py4j==0.10.7
PySocks==1.7.1
pyspark==2.4.5
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
thriftpy2==0.4.10
tweepy==3.8.0
urllib3==1.25.8
Werkzeug==1.0.0
